requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cleanup==8.1.0
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.6
mysqlclient==2.2.4
pillow==10.3.0
PyJWT==2.8.0
sqlparse==0.5.0
tzdata==2024.1
